turn on pipeline writer testing

+++ test/unit/PipelineManagerTest.cpp
@@ -115,10 +115,10 @@ BOOST_AUTO_TEST_CASE(PipelineManagerTest_test4)
 
     mgr.writeWriterPipeline("test.xml");
 
-    bool filesSame = Support::compare_files("test.xml", Support::datapath("pipeline_write_out.xml"));
-    //BOOST_CHECK(filesSame);
+    bool filesSame = Support::compare_text_files("test.xml", Support::datapath("pipeline_write.xml"));
+    BOOST_CHECK(filesSame);
 
-    //if (filesSame)
+    if (filesSame)
     {
         Utils::deleteFile("test.xml");
     }
